# api-gateway/swagger/swagger.yaml
basePath: /
definitions:
  model.CreateIntersectionRequest:
    properties:
      default_parameters:
        $ref: '#/definitions/model.SimulationParameters'
      details:
        properties:
          address:
            example: Corner of Foo and Bar
            type: string
          city:
            example: Pretoria
            type: string
          province:
            example: Gauteng
            type: string
        type: object
      name:
        example: My Intersection
        maxLength: 256
        type: string
      traffic_density:
        example: high
        type: string
    required:
    - default_parameters
    - name
    type: object
  model.CreateIntersectionResponse:
    properties:
      id:
        example: "2"
        type: string
    type: object
  model.Details:
    properties:
      address:
        example: Corner of Foo and Bar
        type: string
      city:
        example: Pretoria
        type: string
      province:
        example: Gauteng
        type: string
    type: object
  model.ErrorResponse:
    properties:
      code:
        example: 404
        type: integer
      message:
        example: resource not found
        type: string
    type: object
...
  model.Intersection:
    properties:
      best_parameters:
        $ref: '#/definitions/model.OptimisationParameters'
      created_at:
        example: "2025-06-24T15:04:05Z"
        type: string
      current_parameters:
        $ref: '#/definitions/model.OptimisationParameters'
      default_parameters:
        $ref: '#/definitions/model.OptimisationParameters'
      details:
        $ref: '#/definitions/model.Details'
      id:
        example: "1"
        type: string
      last_run_at:
        example: "2025-06-24T15:04:05Z"
        type: string
      name:
        example: My Intersection
        type: string
      run_count:
        example: 7
        type: integer
      status:
        example: unoptimised
        type: string
      traffic_density:
        example: high
        type: string
    type: object
  model.Intersections:
    properties:
      intersections:
        items:
          $ref: '#/definitions/model.Intersection'
        type: array
    type: object
...
  model.OptimisationParameters:
    properties:
      optimisation_type:
        example: grid_search
        type: string
      simulation_parameters:
        $ref: '#/definitions/model.SimulationParameters'
    type: object
...
  model.SimulationParameters:
    properties:
      green:
        example: 10
        minimum: 1
        type: integer
      intersection_type:
        example: t-junction
        type: string
      red:
        example: 6
        minimum: 1
        type: integer
      seed:
        example: 3247128304
        type: integer
      speed:
        example: 60
        minimum: 1
        type: integer
      yellow:
        example: 2
        minimum: 1
        type: integer
    required:
    - green
    - intersection_type
    - red
    - seed
    - speed
    - yellow
    type: object
  model.UpdateIntersectionRequest:
    properties:
      details:
        properties:
          address:
            example: Corner of Foo and Bar
            type: string
          city:
            example: Pretoria
            type: string
          province:
            example: Gauteng
            type: string
        type: object
      name:
        example: My Updated Intersection
        type: string
    type: object
...
host: localhost:9090
info:
  contact:
    email: insideinsights2025@gmail.com
    name: Inside Insights Team
    url: https://swagger.io/
  description: This is the API Gateway for the Swift-Signals project, forwarding requests
    to backend gRPC microservices.
  termsOfService: http://example.com/terms/
  title: Authentication API Gateway
  version: "1.0"
paths:
...
  /intersections:
    get:
      consumes:
      - application/json
      description: Retrieves all the intersections associated with the user.
      produces:
      - application/json
      responses:
        "200":
          description: Successful intersections retrieval
          schema:
            $ref: '#/definitions/model.Intersections'
        "401":
          description: 'Unauthorized: Token missing or invalid'
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get All Intersections
      tags:
      - Intersections
    post:
      consumes:
      - application/json
      description: Creates a new intersection with the given arguments
      parameters:
      - description: Intersection information
        in: body
        name: createIntersectionRequest
        required: true
        schema:
          $ref: '#/definitions/model.CreateIntersectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Intersection successfully created
          schema:
            $ref: '#/definitions/model.CreateIntersectionResponse'
        "400":
          description: Invalid request payload or missing fields
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: 'Unauthorized: Token missing or invalid'
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Intersection
      tags:
      - Intersections
  /intersections/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes the intersection with the given ID.
      parameters:
      - description: Intersection ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: 'Unauthorized: Token missing or invalid'
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: 'Not Found: Intersection does not exist'
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Intersection
      tags:
      - Intersections
    get:
      consumes:
      - application/json
      description: Retrieves a single intersection by its unique identifier.
      parameters:
      - description: Intersection ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful intersection retrieval
          schema:
            $ref: '#/definitions/model.Intersection'
        "400":
          description: 'Bad Request: Invalid or missing ID parameter'
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: 'Unauthorized: Token missing or invalid'
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: 'Not Found: Intersection does not exist'
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Intersection by ID
      tags:
      - Intersections
    patch:
      consumes:
      - application/json
      description: Partially updates fields of an existing intersection by ID.
      parameters:
      - description: Intersection ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateIntersectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful update
          schema:
            $ref: '#/definitions/model.Intersection'
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: 'Unauthorized: Token missing or invalid'
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: 'Not Found: Intersection does not exist'
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Intersection
      tags:
      - Intersections
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"